service/requirements.txt
apscheduler
paho-mqtt
RPi.GPIO
